Tuneup number of attempts to fetch blocks.
- Increase `MAX_RECONNECT_ATTEMPTS` to 30

Current time on the fullnode to get new blocks after a session was taking longer than the 5 attempts with the `Math.pow(2)` logic hardcode on backoff retries. So this lead to the indexer keep crashing, which then delay the start of the next block when it is ready since needs to rebuild the code and warmup everything again.